requirements.txt
flask
celery
redis
gunicorn
requests
mmpdb==3.1.3
pyyaml
biopython
py3dmol
numpy
psutil>=5.9.0
python-dotenv
psycopg[binary]
gemmi
rdkit
